requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pyautogui==0.9.54
psutil==5.9.8
pillow==10.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
colorama==0.4.6
websocket-client==1.6.4
